Give own file to ShadowInspectorRow
