+++ Stitch/Graph/Node/Port/View/ShadowInspectorRow.swift
@@ -1,0 +1,34 @@
+//
+//  ShadowInspectorRow.swift
+//  Stitch
+//
+//  Created by Christian J Clampitt on 3/10/25.
+//
+
+import SwiftUI
+
+struct ShadowInputInspectorRow: View {
+    
+    @Environment(\.appTheme) var theme
+    
+    let nodeId: NodeId
+    let propertyIsSelected: Bool
+    
+    var body: some View {
+        HStack {
+            StitchTextView(string: "Shadow",
+                           fontColor: propertyIsSelected ? theme.fontColor : STITCH_FONT_GRAY_COLOR)
+            Spacer()
+        }
+        .overlay {
+            Color.white.opacity(0.001)
+                .onTapGesture {
+                    dispatch(FlyoutToggled(
+                        flyoutInput: SHADOW_FLYOUT_LAYER_INPUT_PROXY,
+                        flyoutNodeId: nodeId,
+                        // No particular field to focus
+                        fieldToFocus: nil))
+                }
+        }
+    }
+}
